fix RunningTime component
- fix design of RunningTime component

- Apply PR comment
