+++ front/src/common/Pathfinding/Pathfinding.tsx
@@ -32,7 +32,6 @@ import {
   getRollingStockID,
   getPathfindingID,
   getGeojson,
-  getMaximumRunTime,
 } from 'reducers/osrdconf/selectors';
 
 import ModalPathJSONDetail from 'applications/operationalStudies/components/ManageTrainSchedule/Itinerary/ModalPathJSONDetail';
@@ -235,7 +234,6 @@ function Pathfinding({ mustUpdate = true, zoomToFeature }: PathfindingProps) {
   const origin = useSelector(getOrigin, isEqual);
   const destination = useSelector(getDestination, isEqual);
   const vias = useSelector(getVias, isEqual);
-  const maximumRunTime = useSelector(getMaximumRunTime, isEqual);
   const rollingStockID = useSelector(getRollingStockID, isEqual);
   const pathfindingID = useSelector(getPathfindingID, isEqual);
   const geojson = useSelector(getGeojson, isEqual);
@@ -377,7 +375,6 @@ function Pathfinding({ mustUpdate = true, zoomToFeature }: PathfindingProps) {
     [!origin, t('origin')],
     [!destination, t('destination')],
     [!rollingStockID, t('rollingstock')],
-    [!maximumRunTime, t('maximumRunTime')],
   ]);
 
   const isPathFindingActive = Object.values(pathfindingState).every(
